style(tests): format code and improve consistency in test cases
- Adjust spacing and formatting in various test files for better readability.
- Ensure consistent use of System.String.Empty in string-related tests.
- Remove unnecessary blank lines and trailing spaces in multiple files.

+++ ConstExpr.Tests/Tests/Hashing/PJWHashTests.cs
@@ -7,21 +7,21 @@ public class PJWHashTests() : BaseTest<Func<string, uint>>(FastMathFlags.All)
 {
 	public override string TestMethod => GetString(str =>
 	{
-		const uint BitsInUnsignedInt = (uint) (sizeof(uint) * 8);
-		const uint ThreeQuarters = (uint) ((BitsInUnsignedInt * 3) / 4);
-		const uint OneEighth = (uint) (BitsInUnsignedInt / 8);
-		const uint HighBits = (uint) (0xFFFFFFFF) << (int) (BitsInUnsignedInt - OneEighth);
+		const uint BitsInUnsignedInt = sizeof(uint) * 8;
+		const uint ThreeQuarters = BitsInUnsignedInt * 3 / 4;
+		const uint OneEighth = BitsInUnsignedInt / 8;
+		const uint HighBits = 0xFFFFFFFF << (int)(BitsInUnsignedInt - OneEighth);
 		uint hash = 0;
 		uint test = 0;
 		uint i = 0;
 
 		for (i = 0; i < str.Length; i++)
 		{
-			hash = (hash << (int) OneEighth) + ((byte) str[(int) i]);
+			hash = (hash << (int)OneEighth) + (byte)str[(int)i];
 
 			if ((test = hash & HighBits) != 0)
 			{
-				hash = ((hash ^ (test >> (int) ThreeQuarters)) & (~HighBits));
+				hash = (hash ^ test >> (int)ThreeQuarters) & ~HighBits;
 			}
 		}
 
@@ -37,7 +37,7 @@ public class PJWHashTests() : BaseTest<Func<string, uint>>(FastMathFlags.All)
 
 			for (var i = 0U; i < str.Length; i++)
 			{
-				hash = hash * 16U + (byte) str[(int) i];
+				hash = hash * 16U + (byte)str[(int)i];
 
 				if ((test = hash & 4026531840U) != 0U)
 					hash = (hash ^ test >> 24) & 268435455U;
